requirements.txt
aiogram==2.25.1
aiohttp
requests
python-dotenv
feedparser
